@@ -45,6 +45,7 @@ class Database:
         self.guilds = Guilds(self.conn)
         self.locks = Locks(self.conn, self)
         self.published_messages = PublishedMessages(self.conn)
+        self.thread_locks = ThreadLocks(self.conn, self)
 
 
 class Infractions(IInfractions):
@@ -87,7 +88,8 @@ class Infractions(IInfractions):
         infraction = None
         try:
             infraction = self.conn.execute(
-                "SELECT * FROM infractions WHERE oid=:id", {"id": infraction_id},
+                "SELECT * FROM infractions WHERE oid=:id",
+                {"id": infraction_id},
             ).fetchone()
         except sqlite3.DatabaseError:
             pass
@@ -189,7 +191,7 @@ class Infractions(IInfractions):
 
     def find_recent_ban_by_id_time_limited(self, user_id, guild_id) -> Infraction:
         infraction = None
-        infraction_on = datetime.utcnow() - timedelta(minutes=1)
+        infraction_on = datetime.now(timezone.utc) - timedelta(minutes=1)
         try:
             infraction = self.conn.execute(
                 "SELECT * FROM infractions WHERE user_id=:user_id "
@@ -209,7 +211,7 @@ class Infractions(IInfractions):
                 DBUser(infraction["moderator_id"], infraction["moderator_name"]),
                 self.db.guilds.find_by_id(guild_id),
                 infraction["reason"],
-                infraction["infraction_on"],
+                infraction["infraction_on"].replace(tzinfo=timezone.utc),
                 InfractionType(infraction["infraction_type"]),
                 self.db.pardons.find_by_id(infraction["oid"]),
                 self.db.published_messages.find_by_id_and_type(
@@ -250,7 +252,7 @@ class Infractions(IInfractions):
                     DBUser(infraction["moderator_id"], infraction["moderator_name"]),
                     self.db.guilds.find_by_id(guild_id),
                     infraction["reason"],
-                    infraction["infraction_on"],
+                    infraction["infraction_on"].replace(tzinfo=timezone.utc),
                     InfractionType(infraction["infraction_type"]),
                     self.db.pardons.find_by_id(infraction["oid"]),
                     self.db.published_messages.find_by_id_and_type(
@@ -292,7 +294,7 @@ class Infractions(IInfractions):
                     DBUser(infraction["moderator_id"], infraction["moderator_name"]),
                     self.db.guilds.find_by_id(guild_id),
                     infraction["reason"],
-                    infraction["infraction_on"],
+                    infraction["infraction_on"].replace(tzinfo=timezone.utc),
                     InfractionType(infraction["infraction_type"]),
                     self.db.pardons.find_by_id(infraction["oid"]),
                     self.db.published_messages.find_by_id_and_type(
@@ -334,7 +336,7 @@ class Infractions(IInfractions):
                     DBUser(infraction["moderator_id"], infraction["moderator_name"]),
                     self.db.guilds.find_by_id(guild_id),
                     infraction["reason"],
-                    infraction["infraction_on"],
+                    infraction["infraction_on"].replace(tzinfo=timezone.utc),
                     InfractionType(infraction["infraction_type"]),
                     self.db.pardons.find_by_id(infraction["oid"]),
                     self.db.published_messages.find_by_id_and_type(
@@ -376,7 +378,7 @@ class Infractions(IInfractions):
                     DBUser(infraction["moderator_id"], infraction["moderator_name"]),
                     self.db.guilds.find_by_id(guild_id),
                     infraction["reason"],
-                    infraction["infraction_on"],
+                    infraction["infraction_on"].replace(tzinfo=timezone.utc),
                     InfractionType(infraction["infraction_type"]),
                     self.db.pardons.find_by_id(infraction["oid"]),
                     self.db.published_messages.find_by_id_and_type(
@@ -447,7 +449,7 @@ class Pardons(IPardons):
                 Pardon(
                     pardon["infraction_id"],
                     DBUser(pardon["moderator_id"], pardon["moderator_name"]),
-                    pardon["pardon_on"],
+                    pardon["pardon_on"].replace(tzinfo=timezone.utc),
                     pardon["reason"],
                 )
                 if pardon
@@ -509,7 +511,7 @@ class Mutes(IMutes):
             return (
                 Mute(
                     self.db.infractions.find_by_id_only(mute["infraction_id"]),
-                    mute["end_time"],
+                    mute["end_time"].replace(tzinfo=timezone.utc),
                     DBUser(mute["user_id"], mute["user_name"]),
                 )
                 if mute
@@ -521,7 +523,7 @@ class Mutes(IMutes):
         try:
             mutes = self.conn.execute(
                 "SELECT * FROM mutes WHERE DATETIME(end_time) < :time",
-                {"time": datetime.utcnow()},
+                {"time": datetime.now(timezone.utc)},
             ).fetchall()
         except sqlite3.DatabaseError:
             pass
@@ -531,7 +533,7 @@ class Mutes(IMutes):
                 objectified_mutes.append(
                     Mute(
                         self.db.infractions.find_by_id_only(mute["infraction_id"]),
-                        mute["end_time"],
+                        mute["end_time"].replace(tzinfo=timezone.utc),
                         DBUser(mute["user_id"], mute["user_name"]),
                     )
                 )
@@ -559,7 +561,7 @@ class Mutes(IMutes):
         try:
             mute = self.conn.execute(
                 "SELECT * FROM mutes WHERE DATETIME(end_time) > :time AND user_id=:user_id",
-                {"time": datetime.utcnow(), "user_id": user_id},
+                {"time": datetime.now(timezone.utc), "user_id": user_id},
             ).fetchone()
         except sqlite3.DatabaseError:
             pass
@@ -567,7 +569,7 @@ class Mutes(IMutes):
             return (
                 Mute(
                     self.db.infractions.find_by_id_only(mute["infraction_id"]),
-                    mute["end_time"],
+                    mute["end_time"].replace(tzinfo=timezone.utc),
                     DBUser(mute["user_id"], mute["user_name"]),
                 )
                 if mute
@@ -675,7 +677,7 @@ class Locks(ILocks):
                     DBUser(lock["moderator_id"], lock["moderator_name"]),
                     self.db.guilds.find_by_id(lock["guild_id"]),
                     lock["reason"],
-                    lock["end_time"],
+                    lock["end_time"].replace(tzinfo=timezone.utc),
                 )
                 if lock
                 else None
@@ -686,7 +688,7 @@ class Locks(ILocks):
         try:
             locks = self.conn.execute(
                 "SELECT * FROM locks WHERE DATETIME(end_time) < :time",
-                {"time": datetime.utcnow()},
+                {"time": datetime.now(timezone.utc)},
             ).fetchall()
         except sqlite3.DatabaseError:
             pass
@@ -702,7 +704,7 @@ class Locks(ILocks):
                         DBUser(lock["moderator_id"], lock["moderator_name"]),
                         self.db.guilds.find_by_id(lock["guild_id"]),
                         lock["reason"],
-                        lock["end_time"],
+                        lock["end_time"].replace(tzinfo=timezone.utc),
                     )
                 )
             return objectified_locks
@@ -753,6 +755,105 @@ class Locks(ILocks):
 
     def delete(self, channel_id: int) -> None:
         self.conn.execute("DELETE FROM locks WHERE channel_id=:id", {"id": channel_id})
+        self.conn.commit()
+
+
+class ThreadLocks(IThreadLocks):
+    def __init__(self, conn: sqlite3.Connection, db: Database):
+        self.conn = conn
+        self.db = db
+
+    def find_by_id(self, channel_id: int) -> Lock:
+        lock = None
+        try:
+            lock = self.conn.execute(
+                "SELECT * FROM thread_locks WHERE channel_id=:id", {"id": channel_id}
+            ).fetchone()
+        except sqlite3.DatabaseError:
+            pass
+        finally:
+            return (
+                ThreadLock(
+                    lock["channel_id"],
+                    DBUser(lock["moderator_id"], lock["moderator_name"]),
+                    self.db.guilds.find_by_id(lock["guild_id"]),
+                    lock["reason"],
+                    lock["end_time"].replace(tzinfo=timezone.utc),
+                )
+                if lock
+                else None
+            )
+
+    def find_expired_locks(self) -> List[ThreadLock]:
+        locks = []
+        try:
+            locks = self.conn.execute(
+                "SELECT * FROM thread_locks WHERE DATETIME(end_time) < :time",
+                {"time": datetime.now(timezone.utc)},
+            ).fetchall()
+        except sqlite3.DatabaseError:
+            pass
+        finally:
+            objectified_locks = []
+            for lock in locks:
+                objectified_locks.append(
+                    ThreadLock(
+                        lock["channel_id"],
+                        DBUser(lock["moderator_id"], lock["moderator_name"]),
+                        self.db.guilds.find_by_id(lock["guild_id"]),
+                        lock["reason"],
+                        lock["end_time"].replace(tzinfo=timezone.utc),
+                    )
+                )
+            return objectified_locks
+
+    def save(self, lock: ThreadLock) -> ThreadLock:
+        retrieved_lock = self.find_by_id(lock.channel_id)
+        if retrieved_lock:
+            try:
+                self.conn.execute(
+                    "UPDATE thread_locks SET "
+                    "moderator_id=:moderator_id,"
+                    "moderator_name=:moderator_name,"
+                    "reason=:reason,"
+                    "end_time=:end_time "
+                    "WHERE channel_id=channel_id",
+                    {
+                        "moderator_id": lock.moderator.id,
+                        "moderator_name": lock.moderator.name,
+                        "reason": lock.reason,
+                        "end_time": lock.end_time,
+                        "channel_id": lock.channel_id,
+                    },
+                )
+                self.conn.commit()
+            except sqlite3.DatabaseError:
+                pass
+        else:
+            try:
+                values = (
+                    lock.channel_id,
+                    lock.moderator.id,
+                    lock.moderator.name,
+                    lock.guild.id,
+                    lock.reason,
+                    lock.end_time,
+                )
+                sql = (
+                    "INSERT INTO thread_locks (channel_id, moderator_id, moderator_name, "
+                    "guild_id, reason, end_time) "
+                    "VALUES(?,?,?,?,?,?)"
+                )
+                self.conn.execute(sql, values)
+                self.conn.commit()
+            except sqlite3.DatabaseError:
+                pass
+        return self.find_by_id(lock.channel_id)
+
+    def delete(self, channel_id: int) -> None:
+        self.conn.execute(
+            "DELETE FROM thread_locks WHERE channel_id=:id", {"id": channel_id}
+        )
         self.conn.commit()
 
 
